# [llvm-profdata] Do not create numerical strings for MD5 function names read from a Sample Profile.