… llvm/include/llvm/ProfileData/SampleProf.h
… llvm/lib/ProfileData/SampleProf.cpp
… llvm/lib/ProfileData/SampleProfReader.cpp
… llvm/lib/ProfileData/SampleProfWriter.cpp
… llvm/lib/Transforms/IPO/SampleContextTracker.cpp
… llvm/lib/Transforms/IPO/SampleProfile.cpp
… llvm/tools/llvm-profdata/llvm-profdata.cpp
… llvm/tools/llvm-profgen/CallContext.h
… llvm/tools/llvm-profgen/ProfileGenerator.cpp
… llvm/tools/llvm-profgen/ProfileGenerator.h
… llvm/tools/llvm-profgen/ProfiledBinary.cpp
… llvm/tools/llvm-profgen/ProfiledBinary.h
… llvm/unittests/ProfileData/SampleProfTest.cpp