@huangjd 2023-09-15
ProfileFuncRef no longer depends on StringRef,