@aeubanks 2023-09-18
a comment on the github workflow, you're not supposed to force push, just create commits on top of the branch, and no need to close the PR unless you're completely abandoning the patch