@huangjd 2023-09-18
> a comment on the github workflow, you're not supposed to force push, just create commits on top of the branch, and no need to close the PR unless you're completely abandoning the patch

When I first created the PR the main branch has changed too much and I had to resolve conflict and update the code in quite a few places, and when I push the rebased branch to my fork repo, this PR now contain several commits that is already checked in, and I don't know how to exclude them (like picking specific commits as diff for review in phab).  What is the correct workflow for this kind of operation?