@huangjd 2023-09-21
Added performance measurement. This patch does significantly reduce profile load time, which is its motivation