@huangjd 2023-09-21
Full profile load is further reduced to 24s after optimizing lazy loading of name table, which was a bottleneck