@david-xl 2023-09-26
The change looks good to me, but wait for other reviewers to chime in as well. Also a good idea to run tests with sanitizers on.