@WenleiHe 2023-09-27
> When testing on an internal large profile (> 1 GB, with more than 10 million functions), the full profile load time is reduced from 28 sec to 25 sec in average, and reading function offset table from 0.78s to 0.7s 

What is the E2E compile time change for the case you measured? Is there any memory footprint change?