@huangjd 2023-10-05
It looks like after using HashKeyMap, the purpose of SymbolMap and GUIDToFuncNameMap can be merged. SymbolMap maps StringRef to Function*, GUIDToFuncNameMap maps ProfileFuncRef's hash code to StringRef, and it's almost exclusively cascaded to SymbolMap lookup, so we can actually combine two maps into one, saving a lot of time.